Remove s3dataset_base.py and refactor S3 datasets
We do not want to expose S3DatasetBase and we face challenges
generating documentation while extending this class.
