+++ s3torchconnector/src/s3torchconnector/__init__.py
@@ -4,7 +4,6 @@ from s3torchconnectorclient import S3Exception
 
 from .s3checkpoint import S3Checkpoint
 from ._s3client import S3Reader, S3Writer
-from .s3dataset_base import S3DatasetBase
 from .s3iterable_dataset import S3IterableDataset
 from .s3map_dataset import S3MapDataset
