+++ s3torchconnector/src/s3torchconnector/s3checkpoint.py
@@ -1,4 +1,7 @@
-from s3torchconnector.s3dataset_base import _parse_s3_uri
+#  Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+#  // SPDX-License-Identifier: BSD
+
+from s3torchconnector._s3dataset_common import parse_s3_uri
 from s3torchconnector._s3client import S3Client, S3Reader, S3Writer
 
 
@@ -8,9 +11,9 @@ class S3Checkpoint:
         self._client = S3Client(region)
 
     def reader(self, s3_uri: str) -> S3Reader:
-        bucket, key = _parse_s3_uri(s3_uri)
+        bucket, key = parse_s3_uri(s3_uri)
         return self._client.get_object(bucket, key)
 
     def writer(self, s3_uri: str) -> S3Writer:
-        bucket, key = _parse_s3_uri(s3_uri)
+        bucket, key = parse_s3_uri(s3_uri)
         return self._client.put_object(bucket, key)
