+++ s3torchconnector/tst/unit/test_s3iterable_dataset.py
@@ -8,7 +8,7 @@ import pytest
 from s3torchconnector import S3IterableDataset
 from s3torchconnector._s3client import S3Reader
 
-from test_s3dataset_base import (
+from test_s3dataset_common import (
     TEST_BUCKET,
     _create_mock_client_with_dummy_objects,
     S3_PREFIX,
